expeditions/requirements.txt
pandas
schema
pyyaml